requirements.txt
requests
pandas
openpyxl
xlsxwriter
python-dotenv
geopy
mvg
ipykernel
jupyter
notebook
gradio